+++ src/parser/inline.ts
@@ -1,10 +1,11 @@
 ﻿import { MarkdownParser } from '../../markdown.d';
 import { union } from '../combinator';
-import { comment } from './inline/comment';
 import { annotation } from './inline/annotation';
+import { authority } from './inline/authority';
 import { link } from './inline/link';
 import { extension } from './inline/extension';
 import { html } from './inline/html';
+import { comment } from './inline/comment';
 import { emphasis } from './inline/emphasis';
 import { strong } from './inline/strong';
 import { code } from './inline/code';
@@ -16,11 +17,12 @@ import { autolink } from './inline/autolink';
 import { text } from './source/text';
 
 export import InlineParser = MarkdownParser.InlineParser;
-export import CommentParser = InlineParser.CommentParser;
 export import AnnotationParser = InlineParser.AnnotationParser;
+export import AuthorityParser = InlineParser.AuthorityParser;
 export import LinkParser = InlineParser.LinkParser;
 export import ExtensionParser = InlineParser.ExtensionParser;
 export import HTMLParser = InlineParser.HTMLParser;
+export import CommentParser = InlineParser.CommentParser;
 export import EmphasisParser = InlineParser.EmphasisParser;
 export import StrongParser = InlineParser.StrongParser;
 export import CodeParser = InlineParser.CodeParser;
@@ -31,11 +33,12 @@ export import HTMLEntityParser = InlineParser.HTMLEntityParser;
 export import AutolinkParser = InlineParser.AutolinkParser;
 
 export const inline: InlineParser = union<InlineParser>([
-  comment,
   annotation,
+  authority,
   extension,
   link,
   html,
+  comment,
   emphasis,
   strong,
   code,