+++ src/parser/inline/authority.ts
@@ -1,9 +1,9 @@
-﻿import { AnnotationParser, inline } from '../inline';
+﻿import { AuthorityParser, inline } from '../inline';
 import { union, some, surround, verify, fmap, build } from '../../combinator';
 import { hasText, hasMedia, hasAnnotationOrAuthority } from '../util';
 import { html } from 'typed-dom';
 
-export const annotation: AnnotationParser = verify(fmap(build(() =>
-  surround('((', some(union([inline]), '))'), '))')),
-  ns => [html('sup', { class: 'annotation' }, ns)]
+export const authority: AuthorityParser = verify(fmap(build(() =>
+  surround('[[', some(union([inline]), ']]'), ']]')),
+  ns => [html('sup', { class: 'authority' }, ns)]
 ), ([el]) => hasText(el) && !hasMedia(el) && !hasAnnotationOrAuthority(el));