+++ src/util/footnote.ts
@@ -1,55 +1,48 @@
 ﻿import { text } from '../parser/block/indexer';
-import { concat } from 'spica/concat';
-import { TypedHTML, html } from 'typed-dom';
+import { TypedHTML, html, define } from 'typed-dom';
 
-const annotation = new WeakMap<HTMLElement, Node[]>();
-const reference = new WeakMap<HTMLElement, HTMLElement>();
-
-export function footnote(source: DocumentFragment | HTMLElement, target: HTMLOListElement): void {
-  const footnotes = new Map<string, HTMLLIElement>();
-  return void TypedHTML.ol([...source.querySelectorAll<HTMLElement>('.annotation')]
-    .reduce<HTMLLIElement[]>((acc, el, i) => {
-      !annotation.has(el) && void annotation.set(el, [...el.childNodes]);
-      reference.has(el) && void reference.get(el)!.remove();
-      void defineTitle(el);
-      const title = el.getAttribute('title')!;
-      void defineId(el, i + 1);
-      const fn = footnotes.get(title);
-      const index = fn
-        ? +fn.getAttribute('id')!.match(/\d+/)![0]
-        : acc.length + 1;
-      const id = fn
-        ? fn.getAttribute('id')!
-        : `footnote:${index}`;
-      assert(id.startsWith('footnote:'));
-      void reference.set(el, el.appendChild(html('a', { href: `#${id}`, rel: 'noopener' }, `[${index}]`)));
-      return fn
-        ? (() => {
-            void fn.lastChild!
-              .appendChild(html('a', { href: `#${el.id}`, rel: 'noopener' }, `[${i + 1}]`));
-            void [...annotation.get(el)!]
-              .forEach(node =>
-                node.parentNode === el &&
-                void el.removeChild(node));
-            return acc;
-          })()
-        : concat(acc, [html('li', { id }, [
-            ...annotation.get(el)!,
-            html('sup', [
-              html('a', { href: `#${el.id}`, rel: 'noopener' }, `[${i + 1}]`),
-            ]),
-          ])].map(el => void footnotes.set(title, el) || el));
-    }, [])
-    .map(el => TypedHTML.li(() => el)),
-    () => target).element;
+export function footnote(source: DocumentFragment | HTMLElement, targets: { annotation: HTMLOListElement; authority: HTMLOListElement; }): void {
+  void annotation(source, targets.annotation);
+  void authority(source, targets.authority);
 }
 
-function defineTitle(target: HTMLElement): void {
-  if (target.hasAttribute('title')) return;
-  void target.setAttribute('title', text(target));
-}
+export const annotation = build('annotation', n => `(${n})`);
+export const authority = build('authority', n => `[${n}]`);
 
-function defineId(target: HTMLElement, index: number): void {
-  if (target.hasAttribute('id')) return;
-  void target.setAttribute('id', `annotation:${index}`);
+function build(category: string, indexer: (index: number) => string): (source: DocumentFragment | HTMLElement, target: HTMLOListElement) => void {
+  assert(category.match(/^[a-z]+$/));
+  const memory = new WeakMap<HTMLElement, Node[]>();
+  return (source: DocumentFragment | HTMLElement, target: HTMLOListElement) => {
+    const definitions = new Map<string, HTMLLIElement>();
+    target.innerHTML = '';
+    void source.querySelectorAll<HTMLElement>(`.${category}`)
+      .forEach((ref, i) => {
+        !memory.has(ref) && void memory.set(ref, [...ref.childNodes]);
+        const refIndex = i + 1;
+        const refId = ref.id || `${category}-ref:${i + 1}`;
+        const title = ref.title || text(ref);
+        const def = definitions.get(title);
+        const defIndex = def
+          ? +def.id.match(/\d+/)![0]
+          : definitions.size + 1;
+        const defId = def
+          ? def.id
+          : `${category}-def:${defIndex}`;
+        ref.innerHTML = '';
+        void define(ref, { id: refId, title: title }, [html('a', { href: `#${defId}`, rel: 'noopener' }, indexer(defIndex))]);
+        return def
+          ? void def.lastChild!.appendChild(html('a', { href: `#${refId}`, rel: 'noopener' }, indexer(refIndex)))
+          : void target.appendChild(
+              TypedHTML.li((f, tag) => {
+                const el = f(tag, { id: defId }, [
+                  ...memory.get(ref)!,
+                  html('sup', [
+                    html('a', { href: `#${refId}`, rel: 'noopener' }, indexer(refIndex)),
+                  ])
+                ]);
+                void definitions.set(title, el);
+                return el;
+              }).element);
+      });
+  };
 }